+++ bank-feeds/src/sdk/models/shared/company.ts
@@ -6,6 +6,15 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { Connection } from "./connection";
 import { Expose, Type } from "class-transformer";
 
+export class GroupReference extends SpeakeasyBase {
+    /**
+     * Unique identifier for the group.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "id" })
+    id?: string;
+}
+
 /**
  * In Codat, a company represents a business sharing access to their data. Each company can have multiple [connections](https://docs.codat.io/bank-feeds-api#/schemas/Connection) to different data sources such as one connection to [Xero](https://docs.codat.io/integrations/accounting/xero/accounting-xero) for accounting data, two connections to [Plaid](https://docs.codat.io/integrations/banking/plaid/banking-plaid) for two bank accounts and a connection to [Zettle](https://docs.codat.io/integrations/commerce/zettle/commerce-zettle) for POS data.
  *
@@ -61,6 +70,14 @@ export class Company extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "description" })
     description?: string;
+
+    /**
+     * An array of groups the company has been assigned to.
+     */
+    @SpeakeasyMetadata({ elemType: GroupReference })
+    @Expose({ name: "groups" })
+    @Type(() => GroupReference)
+    groups?: GroupReference[];
 
     /**
      * Unique identifier for your SMB in Codat.